refactor(McpServer): extract jsonrpc key and version to constants
